+++ agent_cli/agents/_voice_agent_common.py
@@ -9,21 +9,13 @@ import pyperclip
 
 from agent_cli import asr
 from agent_cli.agents._tts_common import handle_tts_playback
-from agent_cli.audio import (
-    input_device,
-    list_input_devices,
-    list_output_devices,
-    output_device,
-)
 from agent_cli.llm import process_and_update_clipboard
 from agent_cli.utils import (
-    print_device_index,
     print_input_panel,
     print_with_style,
 )
 
 if TYPE_CHECKING:
-    import pyaudio
     from rich.live import Live
 
     from agent_cli.agents._config import (
@@ -35,49 +27,6 @@ if TYPE_CHECKING:
     )
 
 LOGGER = logging.getLogger()
-
-
-def setup_devices(
-    p: pyaudio.PyAudio,
-    asr_config: ASRConfig | None,
-    tts_config: TTSConfig | None,
-    quiet: bool,
-) -> tuple[int | None, str | None, int | None] | None:
-    """Handle device listing and setup."""
-    if asr_config and asr_config.list_input_devices:
-        list_input_devices(p)
-        return None
-
-    if tts_config and tts_config.list_output_devices:
-        list_output_devices(p)
-        return None
-
-    # Setup input device
-    input_device_index, input_device_name = input_device(
-        p,
-        asr_config.input_device_name if asr_config else None,
-        asr_config.input_device_index if asr_config else None,
-    )
-    if not quiet:
-        print_device_index(input_device_index, input_device_name)
-
-    # Setup output device for TTS if enabled
-    tts_output_device_index = tts_config.output_device_index if tts_config else None
-    if (
-        tts_config
-        and tts_config.enabled
-        and (tts_config.output_device_name or tts_config.output_device_index)
-    ):
-        tts_output_device_index, tts_output_device_name = output_device(
-            p,
-            tts_config.output_device_name if tts_config else None,
-            tts_config.output_device_index if tts_config else None,
-        )
-        if tts_output_device_index is not None and not quiet:
-            msg = f"🔊 TTS output device [bold yellow]{tts_output_device_index}[/bold yellow] ([italic]{tts_output_device_name}[/italic])"
-            print_with_style(msg)
-
-    return input_device_index, input_device_name, tts_output_device_index
 
 
 async def get_instruction_from_audio(
