+++ agent_cli/agents/speak.py
@@ -13,8 +13,7 @@ import agent_cli.agents._cli_options as opts
 from agent_cli import process_manager
 from agent_cli.agents._config import FileConfig, GeneralConfig, TTSConfig
 from agent_cli.agents._tts_common import handle_tts_playback
-from agent_cli.agents._voice_agent_common import setup_devices
-from agent_cli.audio import pyaudio_context
+from agent_cli.audio import pyaudio_context, setup_devices
 from agent_cli.cli import app, setup_logging
 from agent_cli.utils import (
     get_clipboard_text,
