+++ agent_cli/agents/transcribe.py
@@ -13,8 +13,7 @@ import pyperclip
 import agent_cli.agents._cli_options as opts
 from agent_cli import asr, process_manager
 from agent_cli.agents._config import ASRConfig, GeneralConfig, LLMConfig
-from agent_cli.agents._voice_agent_common import setup_devices
-from agent_cli.audio import pyaudio_context
+from agent_cli.audio import pyaudio_context, setup_devices
 from agent_cli.cli import app, setup_logging
 from agent_cli.llm import process_and_update_clipboard
 from agent_cli.utils import (
